move Task replacement of Placeholders to Tasks::PlaceholderStrategies::HomeworkPersonalized
